+++ index.html
@@ -234,8 +234,8 @@
       </section>
     </template>
 
+    <script src="main.js" type="module"></script>
     <script src="/js/functions.js"></script>
-    <script src="/js/main.js"></script>
 
 </body>
 </html>
